Move ReactWorkTags to the reconciler
