+++ packages/react-dom/src/client/ReactDOMHostConfig.js
@@ -69,7 +69,7 @@ import {
   enableFundamentalAPI,
   enableUseEventAPI,
 } from 'shared/ReactFeatureFlags';
-import {HostComponent} from 'shared/ReactWorkTags';
+import {HostComponent} from 'react-reconciler/src/ReactWorkTags';
 import {
   RESPONDER_EVENT_SYSTEM,
   IS_PASSIVE,
