+++ packages/react-dom/src/events/DeprecatedDOMEventResponderSystem.js
@@ -12,7 +12,11 @@ import {
   PASSIVE_NOT_SUPPORTED,
 } from 'legacy-events/EventSystemFlags';
 import type {AnyNativeEvent} from 'legacy-events/PluginModuleType';
-import {HostComponent, ScopeComponent, HostPortal} from 'shared/ReactWorkTags';
+import {
+  HostComponent,
+  ScopeComponent,
+  HostPortal,
+} from 'react-reconciler/src/ReactWorkTags';
 import type {EventPriority} from 'shared/ReactTypes';
 import type {
   ReactDOMEventResponder,
