+++ packages/react-dom/src/test-utils/ReactTestUtils.js
@@ -14,7 +14,7 @@ import {
   FunctionComponent,
   HostComponent,
   HostText,
-} from 'shared/ReactWorkTags';
+} from 'react-reconciler/src/ReactWorkTags';
 import SyntheticEvent from 'legacy-events/SyntheticEvent';
 import invariant from 'shared/invariant';
 import {ELEMENT_NODE} from '../shared/HTMLNodeType';
