+++ packages/react-reconciler/src/ReactFiberCommitWork.js
@@ -56,7 +56,7 @@ import {
   FundamentalComponent,
   ScopeComponent,
   Block,
-} from 'shared/ReactWorkTags';
+} from './ReactWorkTags';
 import {
   invokeGuardedCallback,
   hasCaughtError,
