+++ packages/react-reconciler/src/ReactFiberContext.js
@@ -12,7 +12,7 @@ import type {StackCursor} from './ReactFiberStack';
 
 import {isFiberMounted} from 'react-reconciler/src/ReactFiberTreeReflection';
 import {disableLegacyContext} from 'shared/ReactFeatureFlags';
-import {ClassComponent, HostRoot} from 'shared/ReactWorkTags';
+import {ClassComponent, HostRoot} from './ReactWorkTags';
 import getComponentName from 'shared/getComponentName';
 import invariant from 'shared/invariant';
 import checkPropTypes from 'shared/checkPropTypes';
