+++ packages/react-test-renderer/src/ReactTestRenderer.js
@@ -40,7 +40,7 @@ import {
   Block,
   IncompleteClassComponent,
   ScopeComponent,
-} from 'shared/ReactWorkTags';
+} from 'react-reconciler/src/ReactWorkTags';
 import invariant from 'shared/invariant';
 import getComponentName from 'shared/getComponentName';
 import ReactVersion from 'shared/ReactVersion';
